services/ordonnances/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.1",
    "beautifulsoup4==4.9.3",
    "fastapi[standard]==0.115.12",
    "opentelemetry-api==1.31.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.31.0",
    "opentelemetry-exporter-otlp-proto-http==1.31.0",
    "opentelemetry-exporter-prometheus==0.52b0",
    "opentelemetry-instrumentation==0.52b0",
    "opentelemetry-instrumentation-fastapi==0.52b0",
    "opentelemetry-instrumentation-grpc==0.52b0",
    "opentelemetry-instrumentation-pymongo==0.52b0",
    "opentelemetry-instrumentation-redis==0.52b0",
    "opentelemetry-instrumentation-requests==0.52b0",
    "opentelemetry-sdk==1.31.0",
    "opentelemetry-semantic-conventions==0.52b0",
    "pika>=1.3.2",
    "pillow==10.3.0",
    "prometheus-client>=0.17.1",
    "psutil>=7.2.1",
    "pyjwt==2.6.0",
    "pymongo==4.6.3",
    "pytesseract==0.3.8",
    "python-dotenv==1.0.1",
    "redis>=7.1.0",
    "reportlab>=4.4.9",
    "requests==2.32.4",
    "schedule==1.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]

[tool.ruff]
target-version = "py312"
line-length = 200

[tool.ruff.lint]
select = [
    "E",   
    "W",   
    "F",   
    "I",   
    "UP",  
    "B",   
    "S",    
    "C4",  
    "DTZ",  
    "T20",  
    "ASYNC",
    "BLE",  
    "PL",   
    "RUF", 
]

ignore = [
    "S101",
    "PLR2004",
]

[tool.ruff.lint.isort]
known-first-party = ["medrax"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.pylint]
max-args = 8

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false